Struct flight_manager_v1::domain::sessions::reserve_tickets::SessionV1

pub struct SessionV1 { /* fields omitted */ }

Won't fix

New design of the payment is recorded in the doc. If I had the time and energy I might refactor the case.

Implement timing facilities of pending reservations.

Implementations

impl SessionV1

pub unsafe fn from_components(
pending_reservations: Box<dyn Storage>,
active_reservations: Box<dyn CreativeStorage>,
users: Arc<dyn Storage>,
items: Arc<dyn Storage>
) -> Self