Trait flight_manager_v1::domain::sessions::reserve_tickets::Session

pub trait Session: Sync {
    pub fn start_reservation(&self, user_id: UserId) -> Result<ReservationId>;

    pub fn add_item(&self, token: ItemToken<'_>) -> Result<()>;

    pub fn remove_item(&self, token: ItemToken<'_>) -> Result<()>;

    pub fn summary(&self, token: UserToken<'_>) -> Result<String>;

    pub fn confirm(&self, token: UserToken<'_>) -> Result<()>;

    pub fn abort(&self, token: UserToken<'_>) -> Result<()>;

    pub fn pay(
        &self, 
        token: UserToken<'_>, 
        p: Box<dyn Payment>
    ) -> Result<Money>;
}

Reserve-Tickets Session.

Required methods

pub fn start_reservation(&self, user_id: UserId) -> Result<ReservationId>

Start a new reservation for the user. Returns a unique id for identifying the reservation.

Temporary Reservation

This call starts a temporary reservation. Until confirmed, the reservation may not be stored persistently.

Error

if the user is not found.

Implementation

Specifically, the method collaborates with others in this way:

check with the user storage to see if the user id is valid.

ask a reservation factory to produce a new reservation for the user.

store the new reservation to the active-reservation storage.

return the id of the reservation.

pub fn add_item(&self, token: ItemToken<'_>) -> Result<()>

Adds a reservable item to the reservation list.

Will exclusively occupy the item.

Error

if any of the user, the reservation, or the item is not found;

if the reservation is already confirmed or not valid (e.g. aborted).

if the item is occupied, which is possible with multiple users accessing the system.

pub fn remove_item(&self, token: ItemToken<'_>) -> Result<()>

Removes an item from the list.

Similar to add.

pub fn summary(&self, token: UserToken<'_>) -> Result<String>

Gets a summary of the current state of the reservation.

Return

a serialized form of the reservation.

Error

if the user id is not conformant.

AFA (Jun. 11 '20), I'm using the YAML form instead of the more famous JSON, since YAML is more concise and readable than JSON.
I've switched back to JSON since it has the best library support.

pub fn confirm(&self, token: UserToken<'_>) -> Result<()>

Confirms an reservation.

Calling this function will terminate the modifying process of the reservation (i.e. add or remove). Any subsequent calls to those functions or this one will result in an error.

Set due time (todo)

When confirming a reservation, the due time of it is updated.

Error

if the reservation is not active.

pub fn abort(&self, token: UserToken<'_>) -> Result<()>

Aborts a reservation that's not paid yet.

The aborted reservation will not be stored in the system.

Error

if the reservation is not active.

pub fn pay(&self, token: UserToken<'_>, p: Box<dyn Payment>) -> Result<Money>

Pays for a reservation.

Expects

the reservation is confirmed.

Ensures

Unpaid reservations that exceed the timeout will be discarded, and will not be recorded in the users' profile. (won't do)

Returns

money actually paid.

Error

if the payment failed

if the reservation is already paid.

if the reservation is not found.

Loading content...