Trait flight_manager_v1::domain::storage::reservations::Storage

pub trait Storage: Sync {
    pub fn add_item(&self, tok: ItemToken<'_>) -> Result<()>;

    pub fn remove_item(&self, tok: ItemToken<'_>) -> Result<()>;

    pub fn summary(&self, tok: UserToken<'_>) -> Result<String>;

    pub fn process(
        &self, 
        tok: UserToken<'_>, 
        op: Box<dyn FnOnce(&Reservation) -> Result<Money>>
    ) -> Result<Money>;

    pub unsafe fn store(&self, r: Reservation);

    pub unsafe fn extract(&self, tok: UserToken<'_>) -> Result<Reservation>;

    pub fn transfer_to(
        &self, 
        tok: UserToken<'_>, 
        other: &dyn Storage
    ) -> Result<()> { ... }
}

Reservation Storage.

Different from other storages, such as UserStorage, which might be singletons, there may be multiple instances of reservation storages present. For example, we may need temporary storage for active reservations which are not confirmed.

Required methods

pub fn add_item(&self, tok: ItemToken<'_>) -> Result<()>

Add the item to the reservation list, does not check if the item exists (it can't anyway).

Error

if user id not conformant

if reservation not found.

pub fn remove_item(&self, tok: ItemToken<'_>) -> Result<()>

pub fn summary(&self, tok: UserToken<'_>) -> Result<String>

Generate a summary of the reservation.

Error

similar.

pub fn process(
&self,
tok: UserToken<'_>,
op: Box<dyn FnOnce(&Reservation) -> Result<Money>>
) -> Result<Money>

Performs an action on a reservation.

pub unsafe fn store(&self, r: Reservation)

Stores a reservation into the storage.

Panic

if cannot store, e.g. memory used up.

Unsafe

Reservations should stay in storages. This function is intended for inter-storage transferring, not for direct use.

You can call CreativeStorage::new_reservation if you want to create a reservation.

pub unsafe fn extract(&self, tok: UserToken<'_>) -> Result<Reservation>

Extract a reservation from the storage.

Error

similar.

Unsafe

Reservations should stay in storages. This function is intended for inter-storage transferring. You seldom have to delete a reservation, right?

Loading content...
Provided methods

pub fn transfer_to(&self, tok: UserToken<'_>, other: &dyn Storage) -> Result<()>

Transfer a reservation in this storage to another.

Error

similar

Loading content...