Trait flight_manager_v1::domain::storage::users::Storage

pub trait Storage: Sync + Send {
    pub fn user_exists(&self, user_id: &UserId) -> bool;

    pub fn link(&self, token: UserToken<'_>);

    pub fn undone_reservations(
        &self, 
        user_id: &UserId
    ) -> Result<Vec<ReservationId>>;

    pub fn done_reservations(
        &self, 
        user_id: &UserId
    ) -> Result<Vec<ReservationId>>;

    pub fn withdraw(&self, token: UserToken<'_>) -> Result<()>;
}

User Storage

Reservations Stored in User Storage

Since reservations are involved with money, and refunding is required, we need to record that "who did what reservations". Reservations are no longer stored as an independent state as in previous tries, since it makes little sense for a reservation to live with no owner, and we can easily get around the limitation by providing a void owner.

Nonetheless, the program still has some reservation storages. But they are mainly used to store temporary reservations (undone, unpaid etc).

My brain must be stuck with shit when I wrote the previous paragraphs. How on earth could I ended up designing a user storage that is so coupled with reservations?

Someone help kill the me of yesterday!

Required methods

pub fn user_exists(&self, user_id: &UserId) -> bool

pub fn link(&self, token: UserToken<'_>)

Links a reservation to the user's profile.

Panic

if the user is not in the storage.

When we decide to call this function, it means that we already have a reservation linked with a certain user in the storage, and we only have to inform the storage of it now. So, it is unlikely that the user could not be found.

pub fn undone_reservations(
&self,
user_id: &UserId
) -> Result<Vec<ReservationId>>

Generate a list of reservation related with the user.

pub fn done_reservations(&self, user_id: &UserId) -> Result<Vec<ReservationId>>

Similar to [undone_reservations]

pub fn withdraw(&self, token: UserToken<'_>) -> Result<()>

Loading content...